Calculate force using radial equation =====================================
findforce.m -- find the force
solvemodel.m -- changed time span and IC's to run model only for last push. This is done as the radius function isn't good for small theta as the limits have had to be streached
radius.m -- changed theta_lim and made the fudge factor between ss and tanh() parts calculated rather than guessed as discontinuities spank the findforce function.
LvFRun.m -- simple wrapper to run other bits. The outputed graphs are scaled so that force, radius and theta can be displayed on the same axis.